refactor(Car): field명 position으로 변경
